two more construction bugs to fix
