+++ honeybee_energy/construction/window.py
@@ -246,23 +246,6 @@ window-calculation-module.html#step-4.-determine-layer-solar-transmittance
                 count += 1
         return count
 
-    @property
-    def user_data(self):
-        """Get or set an optional dictionary for additional meta data for this object.
-
-        This will be None until it has been set. All keys and values of this
-        dictionary should be of a standard Python type to ensure correct
-        serialization of the object to/from JSON (eg. str, float, int, list, dict)
-        """
-        return self._user_data
-    
-    @user_data.setter
-    def user_data(self, value):
-        if value is not None:
-            assert isinstance(value, dict), 'Expected dictionary for honeybee_energy' \
-                'object user_data. Got {}.'.format(type(value))
-        self._user_data = value    
-
     def temperature_profile(self, outside_temperature=-18, inside_temperature=21,
                             wind_speed=6.7, height=1.0, angle=90.0, pressure=101325):
         """Get a list of temperatures at each material boundary across the construction.
